+++ src/validators/shared.ts
@@ -1,8 +1,4 @@
 import { z } from 'zod';
-import { CompletionTokenDetailsSchema, TokenUsageSchema } from '../types/shared';
-
-// Export the schemas to maintain backward compatibility
-export { CompletionTokenDetailsSchema, TokenUsageSchema };
 
 export const NunjucksFilterMapSchema = z.record(
   z.string(),
